backend/requirements.txt
# ====================================
# Monitor de Orçamento Público Municipal
# Backend Requirements
# ====================================

# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6  # Para upload de arquivos

# Pydantic para validação
pydantic==2.5.3
pydantic-settings==2.1.0  # Para carregar .env

# Database
sqlalchemy==2.0.25
alembic==1.13.1  # Migrations
aiosqlite==0.19.0  # Async SQLite

# Google Gemini AI
google-generativeai>=0.8.6

# ChromaDB - Vector Database (0.5+ suporta API v2)
chromadb==0.5.5

# PDF Processing
pypdf==3.17.4
pdfplumber==0.10.4
PyMuPDF==1.23.21  # fitz

# LangChain para chunking e embeddings
langchain==0.1.4
langchain-community==0.0.16
langchain-google-genai==0.0.6

# HTTP Client para Portal da Transparência
httpx>=0.27.0
requests==2.31.0

# Redis para cache e filas
redis==5.0.1
redis[hiredis]==5.0.1  # Com parser C mais rápido

# Celery para tarefas assíncronas
celery==5.3.6

# Data Processing (Fase 2: Schema Discovery)
pandas==2.1.4  # Para análise de CSV
unidecode==1.3.7  # Para normalização de texto

# Utilities
python-dotenv==1.0.0  # Carregar .env
python-jose[cryptography]==3.3.0  # JWT (futuro)
passlib[bcrypt]==1.7.4  # Hash de senhas (futuro)
python-dateutil==2.8.2
python-magic==0.4.27  # Detectar tipo de arquivo

# Logging e Monitoring
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1  # Formatação
flake8==7.0.0  # Linting
mypy==1.8.0  # Type checking
ipython==8.20.0  # Shell interativo
